@@ -65,20 +65,22 @@
     "ie 11"
   ],
   "dependencies": {
-    "@angular/animations": "6.0.2",
-    "@angular/cdk": "6.0.2",
-    "@angular/common": "6.0.2",
-    "@angular/compiler": "6.0.2",
-    "@angular/core": "6.0.2",
-    "@angular/forms": "6.0.2",
-    "@angular/http": "6.0.2",
-    "@angular/material": "^6.4.6",
-    "@angular/platform-browser": "6.0.2",
-    "@angular/platform-browser-dynamic": "6.0.2",
-    "@angular/platform-server": "6.0.2",
-    "@angular/router": "6.0.2",
-    "@covalent/core": "2.0.0-beta.2",
+    "@angular/animations": "6.1.10",
+    "@angular/cdk": "6.4.7",
+    "@angular/common": "6.1.10",
+    "@angular/compiler": "6.1.10",
+    "@angular/core": "6.1.10",
+    "@angular/forms": "6.1.10",
+    "@angular/http": "6.1.10",
+    "@angular/material": "^6.1.10",
+    "@angular/platform-browser": "6.1.10",
+    "@angular/platform-browser-dynamic": "6.1.10",
+    "@angular/platform-server": "6.1.10",
+    "@angular/router": "6.1.10",
+    "@covalent/core": "2.0.0-beta.3",
     "@covalent/tools": "2.0.0-beta.2",
+    "@ngx-translate/core": "10.0.2",
+    "@ngx-translate/http-loader": "3.0.1",
     "@covalent/highlight": "^2.0.0-beta.2",
     "@covalent/markdown": "^2.0.0-beta.2",
     "classlist.js": "^1.1.20150312",
@@ -88,14 +90,15 @@
     "hammerjs": "^2.0.8",
     "highlight.js": "9.13.0",
     "rxjs": "6.0.0",
+    "showdown": "1.6.4",
     "tslib": "^1.9.0",
     "web-animations-js": "2.3.1",
     "zone.js": "^0.8.24"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.1",
-    "@angular/cli": "6.0.1",
-    "@angular/compiler-cli": "6.0.2",
+    "@angular-devkit/build-angular": "~0.8.5",
+    "@angular/cli": "6.2.5",
+    "@angular/compiler-cli": "6.1.10",
     "@types/fs-extra": "^4.0.0",
     "@types/hammerjs": "^2.0.30",
     "@types/jasmine": "~2.8.6",
