@@ -1,15 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 
 @Component({
   selector: 'app-tree-docs',
   templateUrl: './tree-docs.component.html',
-  styleUrls: ['./tree-docs.component.scss']
+  styleUrls: ['./tree-docs.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  preserveWhitespaces: true,
 })
-export class TreeDocsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
+export class TreeDocsComponent {
 
 }
