@@ -5,6 +5,7 @@ export { TdChartYAxisComponent } from './axis/y-axis.component';
 export { TdChartXAxisComponent } from './axis/x-axis.component';
 export { TdChartTooltipComponent } from './tooltip/tooltip.component';
 export { TdChartOptionsService, CHART_PROVIDER } from './chart.service';
+export { ChartRenderDirective } from './directives/chart-render.directive';
 export * from './utils';
 export {
   ITdAxisLine,
@@ -23,7 +24,7 @@ export {
 } from './axis/axis.interface';
 
 export {
-    TdAlign,
+  TdAlign,
   TdFontStyle,
   TdFontWeight,
   TdVerticalAlign,
@@ -48,6 +49,7 @@ export {
   ITdAxisPointerLabel,
   ITdBaseStyle,
   ITdTooltipAxisPointer,
+  ITdCharMarkEvent,
   TdAxisPointerAxis,
 } from './base.types';
 
